project/requirements.txt
flask
opencv-python-headless
deepface
tensorflow
numpy
pillow
gunicorn
